scoreDiff copy
